Fixes pylint errors.
